# go-server/go/model_expenses_response.go
/*
 * Money Tracker API
 *
 * This is a money tracker api
 *
 * API version: 1.0.0
 * Contact: kawakawaryuryu@hotmail.co.jp
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type ExpensesResponse struct {

	// 支出一覧リスト
	Expenses []ExpenseResponse `json:"expenses,omitempty"`
}