# go-server/go/model_not_found.go
/*
 * Money Tracker API
 *
 * This is a money tracker api
 *
 * API version: 1.0.0
 * Contact: kawakawaryuryu@hotmail.co.jp
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type NotFound struct {

	// エラーコード
	Code string `json:"code,omitempty"`

	// エラーメッセージ
	Message string `json:"message,omitempty"`
}